Question about this package
Hi @tonyfreeman,

Sorry to abuse a PR this way but it's the only way I could find to get in touch.  I was wondering about the suitability of this code for something I'm working on.  It's a little language interpreter written in "trampolined continuation-passing style.”  The key point is that the interpreter's user gets control back after each unit of work.  CPS can be awkward, and it would be much cleaner to intersperse `yield`s in a regular eager recursive interpreter implementation, run as a fiber, and just keep poking the fiber to move to the next step.  But the examples I see here don't make it clear whether I can build such a structure.  `Channel` looks kind of close; I wonder if you can offer any insight?  Thanks!